Fix TT bug - not storing full zobrist jey so was getting lots of false positives, also was not validating the best move. Fix Quiescence search, consider quiet moves if in check, and return a checkmate score if no move is legal. Add a minimum weight to opening book moves to avoid any low-quality openings. Update the game phase calculation to be more accurate and to be a normalised float. Reset jupiter's internal metrics each time init is called on it.

+++ lib/src/evaluator.h
@@ -14,7 +14,7 @@ public:
     Evaluator(const AttackTable& attackTable)
         : m_AttackTable{std::forward<const AttackTable>(attackTable)} {}
     int64_t Evaluate(const BoardState& state) const;
-    uint8_t GamePhase(const BoardState& state) const;
+    float GamePhase(const BoardState& state) const;
     int64_t SEE(const BoardState& state, Move move) const;
 
 private:
